requirements.txt
transformers==4.37.2
peft==0.9.0
bitsandbytes==0.42.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pandas==2.2.1
torch==2.2.1
streamlit==1.32.2
cryptography==42.0.5
google-generativeai==0.4.1
scikit-learn==1.4.1
datasets==2.18.0
